MINTCRM-349: Integrate UI with API for double opt in settings            - redirect url validation

+++ app/Admin/API/Controllers/OptinSettingController.php
@@ -45,9 +45,21 @@ class OptinSettingController extends SettingBaseController {
 
         // Get values from API
         $params = MRM_Common::get_api_params_values( $request );
-        error_log(print_r($params, 1));
+        
         if( array_key_exists( 'optin', $params ) ){
             $setting_value = isset( $params['optin'] ) ? $params['optin'] : [];
+
+            $confirmation_type = isset( $setting_value['confirmation_type'] ) ? $setting_value['confirmation_type'] : "";
+            // URL validation
+            $url = isset( $setting_value['url'] ) ? $setting_value['url'] : "";
+            if( "redirect" == $confirmation_type && filter_var($url, FILTER_VALIDATE_URL) === FALSE ){
+                return $this->get_error_response(__( 'Redirect URL is not valid', 'mrm' ));
+            }
+
+            if( "redirect" == $confirmation_type && empty( $url ) )	{
+                return $this->get_error_response(__( 'Redirect URL is missing', 'mrm' ));
+            }
+
             update_option('_mrm_optin_settings',  $setting_value);
             return $this->get_success_response( __("Double opt-in settings have been successfully saved.", "mrm") );
         }
